requirements.txt
pdf2image==1.17.0
pillow==11.2.1
PyPDF2==3.0.1
pytesseract==0.3.13
reportlab==4.4.0
spacy==3.8.5
streamlit==1.45.1
